Create handle Api request file
Relates #27
