+++ utils/helpers/handleElements.js
@@ -7,3 +7,7 @@ export const createElement = (tagName, attributes = {}) => {
 
   return element;
 };
+
+export const appendElements = (parent, elements) => {
+  elements.forEach((element) => parent.appendChild(element));
+};
